Make rendering faster
draw.Draw is very slow when the image is not an RGBA image. This updates
the call to make 2 passes to render: one which uses a DynamicMemoryDrawer
in order to measure the size of the image, without drawing into it, and
one which creates an RGBA image of those dimensions and draws into it.

This results in about a 2x speed increase, despite the doubling of overhead.