Hack to Improve Rendering Speed for Large Pages
It now makes a pass to render only the first page, aborting early,
and then another pass in a goroutine to render the entire page,
using a channel to abort when another page is loaded.

The better way to do this would be to use a gorouting to progressively
notify the viewport when more of the page is rendered, but I couldn't
get that to work properly.. so for now, this allows large pages to
render something relatively quickly, even if you need to wait before
you can scroll.