Refactored and optimized Layout and Drawing code.
Cleanly separated the layout from the drawing code. This results in
the drawing needing to not duplicate any work, and results in about
a 2x speed increase, as well as making it easier to implement things
like input elements which may need to only update a portion of the
image at a time in the future.