Rework floating box calculation algorithm to improve standards compliance
Previously, we would try to reflow all the text after every float was positioned.
This was neither efficient nor effective. We now determine if the float overlaps
with any text that's already been rendered, and if so either move the rendered
text (if there's room) or move the new float box down a line (if there's not).