Handle case where floats overlap margins better
Adjust the coordinate system of the float to the child on the X axis
as well as the Y axis when children inherit floats, and be more rigorous
with differentiating between the widht and the max x of the float. This
helps with the case where a float may overlap a margin.